requirements.txt
pyTelegramBotAPI==4.9.0
Pillow==9.3